refactor: Modularize workflow components and enhance error handling
